+++ Documentation/Changelog.md
@@ -14,6 +14,7 @@ Release date: not released yet
 - Added `Gapotchenko.FX.IOptional` interface for `Gapotchenko.FX.Optional<T>` type to allow an untyped value introspection
 - Implemented memory span-based operations for `Gapotchenko.FX.IO.FragmentedMemoryStream`
 - Polyfills:
+    - Added polyfill for `System.Runtime.CompilerServices.OverloadResolutionPriorityAttribute` type
     - Added `Zip(first, second)` and `Zip(first, second, third)` polyfill methods for `IEnumerable<T>` type
 - Fixed issues:
     - Fixed issue with nested `Gapotchenko.FX.Optional<T>` values that could occur during implicit value conversion